Research/ascii_art/pyproject.toml
[tool.poetry]
name = "ascii-art"
version = "0.1.0"
description = ""
authors = ["Antonio Zea Jr <zeajantonio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.0.0"
structlog = "^23.1.0"
numpy = "^1.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
